Update Telegram user model and SQL migration for referral tracking
- Modified the NewUser interface to include new fields: reinvest, reinvest_setup, aicore_balance, wallet_balance, level, and paid_referrals, with appropriate default values.
- Updated the SQL migration to change the parameter type of the count_user_referrals function from INTEGER to BIGINT, ensuring compatibility with larger telegram_id values.

+++ supabase/migrations/20240328000000_add_referral_count_function.sql
@@ -1,5 +1,5 @@
 -- First, create the counting function
-CREATE OR REPLACE FUNCTION count_user_referrals(p_telegram_id INTEGER)
+CREATE OR REPLACE FUNCTION count_user_referrals(p_telegram_id BIGINT)
 RETURNS INTEGER AS $$
 DECLARE
     referral_count INTEGER;
